+++ sketch_map_tool/definitions.py
@@ -1,10 +1,50 @@
+import json
+import os
+from pathlib import Path
 from typing import Literal
 
-from sketch_map_tool.models import PaperFormat
+from werkzeug.utils import secure_filename
+
+from sketch_map_tool.config import get_config_value
+from sketch_map_tool.models import LiteratureReference, PaperFormat
 
 CREATE_TYPES = Literal["quality-report", "sketch-map"]
 DIGITIZE_TYPES = Literal["digitized-data"]
 ALLOWED_TYPES = Literal[CREATE_TYPES, DIGITIZE_TYPES]
+
+
+def get_literature_references() -> list[LiteratureReference]:
+    """Read a list of literature references from JSON stored on disk.
+
+    For image source either a web URL or a filename of a file in the publications folder
+    is expected.
+    """
+    p = Path(get_config_value("data-dir")) / "literature.json"
+    with open(p, "r") as f:
+        raw = json.load(f)
+
+    def create_literature_reference(element) -> LiteratureReference:
+        # if image is stored on disk
+        img_src = element.get("imgSrc", None)
+        if img_src is not None and not img_src.strip().startswith("http"):
+            img_src = os.path.join(
+                "/static",
+                "assets",
+                "images",
+                "about",
+                "publications",
+                secure_filename(img_src),
+            )
+        return LiteratureReference(
+            element["citation"],
+            img_src,
+            element.get("url", None),
+        )
+
+    return [create_literature_reference(element) for element in raw]
+
+
+LITERATURE_REFERENCES = get_literature_references()
 
 
 A0 = PaperFormat(